added some types to it's own type file and moved some ui components to seperate folder
