+++ client/src/components/LogoutButton.tsx
@@ -1,7 +1,7 @@
 import { useAccount } from '../contexts/AccountContext'
 import { useTransactions } from '../contexts/TransactionsContext'
 import { useUser } from '../contexts/UserContext'
-import { Button } from './Button'
+import { Button } from './ui/Button'
 import { IconLogout } from '@tabler/icons-react'
 
 export const LogoutButton = ({ className }: { className?: string }) => {
@@ -21,8 +21,12 @@ export const LogoutButton = ({ className }: { className?: string }) => {
 	}
 
 	return (
-		<Button className={`${className}`} variant='dashboard' onClick={handleLogout}>
-			<IconLogout size={18}/>
+		<Button
+			className={`${className}`}
+			variant="dashboard"
+			onClick={handleLogout}
+		>
+			<IconLogout size={18} />
 			Log Out
 		</Button>
 	)
