+++ client/src/components/ReceiptCapture.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useRef, useState } from 'react'
-import { Button } from './Button'
+import { Button } from './ui/Button'
 
 const ReceiptCapture = ({
 	onCapture,
