+++ client/src/components/ui/Input.tsx
@@ -1,9 +1,6 @@
-type Input = React.InputHTMLAttributes<HTMLInputElement> & {
-	className?: string
-	variant?: string
-}
+import type { InputType } from '../../types/types'
 
-export const Input = ({ className, variant, ...props }: Input) => {
+export const Input = ({ className, variant, ...props }: InputType) => {
 	const base = `border rounded-lg mb-5 p-2 pl-4 border-gray-400 focus:outline-4 outline-gray-300 `
 
 	return <input className={`${base} ${className}`} {...props}></input>
