+++ client/src/components/ui/Label.tsx
@@ -1,0 +1,10 @@
+import type { LabelType } from '../../types/types'
+
+export const Label = ({ children, className, ...props }: LabelType) => {
+	const base = 'mb-1 font-medium'
+	return (
+		<label className={`${base} ${className}`} {...props}>
+			{children}
+		</label>
+	)
+}
