+++ client/src/contexts/AccountContext.tsx
@@ -1,29 +1,5 @@
-import {
-	createContext,
-	useContext,
-	useState,
-	type Dispatch,
-	type SetStateAction,
-} from 'react'
-
-type Account = {
-	id: string | number
-	account_name: string
-	bank_name: string
-	user_id: string
-	balances: {
-		available: number
-		current: number
-		currency_code: string
-	}
-	subtype: string
-	type: string
-}
-
-type AccountContextType = {
-	accounts: Account[]
-	setAccounts: Dispatch<SetStateAction<Account[]>>
-}
+import { createContext, useContext, useState } from 'react'
+import type { Account, AccountContextType } from '../types/types'
 
 const defaultAccountContextType: AccountContextType = {
 	accounts: [],
