+++ client/src/pages/LandingPage.tsx
@@ -1,5 +1,5 @@
 import { Link } from 'react-router'
-import { Button } from '../components/Button'
+import { Button } from '../components/ui/Button'
 
 const LandingPage = () => {
 	return (
